# kpi-enterprise/api/openapi.v1.yaml
openapi: 3.1.0
info:
  title: KPI Enterprise Mining API
  version: 1.0.0
  description: |
    Public API v1 für KPI Enterprise Mining. Kanonische Quelle: `ARCHITECTURE-v2.md §3.4`.

    **Auth:** Keycloak-JWT (Bearer). Pflichtclaim `tenant_id`. Der API-Gateway
    (`api-gateway:4001`) verifiziert den JWT, extrahiert den Tenant und injiziert
    `app.tenant` in alle Datenbankverbindungen (Postgres RLS, Trino, ClickHouse).

    **Tenant-Filter:** Alle Endpunkte mit `tenant_id`-Filter sind automatisch auf
    den JWT-Tenant beschränkt. Cross-Tenant-Zugriffe werden vom Gateway abgewiesen.

    **Versionierung:** SemVer. Breaking Changes nur als neue Major-Version (`/v2`).

    **Idempotenz:** Schreibende Endpunkte unterstützen `Idempotency-Key` Header.

  contact:
    name: KPI Enterprise Mining Platform
    email: platform@medialine.app
  license:
    name: Proprietary
servers:
  - url: https://api.ben-e-fit.ai/v1
    description: Production (Canonical)
  - url: https://api.medialine.app/v1
    description: Operator/Showcase
  - url: https://api.ki-guru.com/v1
    description: Sister-Tenant
...
security:
  - bearerAuth: []
paths:
...
  /tenants:
    get:
      tags: [tenants]
      summary: Liste aller Tenants (Platform-Admin)
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Tenant' }
    post:
      tags: [tenants]
      summary: Tenant anlegen (Onboarding-Wizard, Platform-Admin)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TenantCreate' }
      responses:
        '201':
          description: created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Tenant' }
  /tenants/{tenantId}:
    parameters:
      - $ref: '#/components/parameters/TenantId'
    get:
      tags: [tenants]
      summary: Tenant-Detail
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Tenant' }
    patch:
      tags: [tenants]
      summary: Tenant-Update
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TenantUpdate' }
      responses:
        '200': { description: updated }
    delete:
      tags: [tenants]
      summary: Tenant-Offboarding (synchronisiert mit Keycloak/MinIO/CH/PG)
      responses:
        '202': { description: deletion enqueued }
  /sources:
    get:
      tags: [sources]
      summary: Datenquellen
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Source' }
    post:
      tags: [sources]
      summary: Datenquelle anlegen
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SourceCreate' }
      responses:
        '201':
          description: created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Source' }
  /sources/{sourceId}/sync:
    parameters:
      - name: sourceId
        in: path
        required: true
        schema: { type: string }
    post:
      tags: [sources]
      summary: Manueller Sync triggern
      responses:
        '202':
          description: queued
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AgentRunRef' }
...
  /datasets/{datasetId}/profile:
    parameters:
      - name: datasetId
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [datasets]
      summary: Spalten-Profil
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DatasetProfile' }
  /metric-definitions:
    get:
      tags: [metrics]
      summary: Liste der Metric Contracts (aktive Versionen)
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/MetricDefinition' }
    post:
      tags: [metrics]
      summary: Metric Contract neu anlegen
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MetricDefinitionCreate' }
      responses:
        '201':
          description: created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MetricDefinition' }
  /metric-definitions/{metricId}/versions:
    parameters:
      - name: metricId
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [metrics]
      summary: Vertrags-Versionen
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/MetricDefinition' }
  /kpi-candidates:
    get:
      tags: [metrics]
      summary: KPI-Kandidaten (Discovery-Output)
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/KpiCandidate' }
    post:
      tags: [metrics]
      summary: Discovery-Run anstoßen
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DiscoveryRunRequest' }
      responses:
        '202':
          description: queued
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AgentRunRef' }
  /kpi-candidates/{candidateId}/promote:
    parameters:
      - name: candidateId
        in: path
        required: true
        schema: { type: string }
    post:
      tags: [metrics]
      summary: Kandidat zu Metric Contract promoten (HITL Approval Pflicht)
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PromoteRequest' }
      responses:
        '201':
          description: promoted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MetricDefinition' }
  /kpi-observations:
    get:
      tags: [kpis]
      summary: Zeitreihe einer Metric
      parameters:
        - name: metric_id
          in: query
          required: true
          schema: { type: string }
        - name: from
          in: query
          schema: { type: string, format: date-time }
        - name: to
          in: query
          schema: { type: string, format: date-time }
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/KpiObservation' }
  /driver-trees:
    get:
      tags: [trees]
      summary: Driver-Trees auflisten
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/DriverTree' }
    post:
      tags: [trees]
      summary: Driver-Tree anlegen
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DriverTreeCreate' }
      responses:
        '201':
          description: created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DriverTree' }
  /driver-trees/{treeId}/nodes:
    parameters:
      - name: treeId
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [trees]
      summary: Knoten
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/DriverNode' }
    post:
      tags: [trees]
      summary: Knoten hinzufügen
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DriverNode' }
      responses:
        '201': { description: created }
  /driver-trees/{treeId}/edges:
    parameters:
      - name: treeId
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [trees]
      summary: Kanten
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/DriverEdge' }
    post:
      tags: [trees]
      summary: Kante hinzufügen
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DriverEdge' }
      responses:
        '201': { description: created }
  /okr-links:
    get:
      tags: [okrs]
      summary: KPI ↔ OKR-Verknüpfungen
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/OkrLink' }
    post:
      tags: [okrs]
      summary: Verknüpfung anlegen
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OkrLink' }
      responses:
        '201': { description: created }
  /insights:
    get:
      tags: [insights]
      summary: Anomaly + RCA Outputs
      parameters:
        - name: severity
          in: query
          schema:
            type: string
            enum: [info, warn, critical]
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Insight' }
  /briefings:
    get:
      tags: [briefings]
      summary: Briefing-Runs
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Briefing' }
    post:
      tags: [briefings]
      summary: Briefing erzeugen
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BriefingCreate' }
      responses:
        '202':
          description: queued
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AgentRunRef' }
  /briefings/{briefingId}:
    parameters:
      - name: briefingId
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [briefings]
      summary: Briefing-Detail (Markdown + PDF Link)
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Briefing' }
  /agent-runs:
    get:
      tags: [agents]
      summary: Agent-Run-Liste (Audit)
      parameters:
        - name: agent
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema:
            type: string
            enum: [queued, running, succeeded, failed, blocked]
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AgentRun' }
  /agent-runs/{runId}:
    parameters:
      - name: runId
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [agents]
      summary: Agent-Run-Detail (Trace + Evidence)
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AgentRun' }
  /evidence/{evidenceId}:
    parameters:
      - name: evidenceId
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [agents]
      summary: Evidence-Bundle (signed URL → MinIO)
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EvidenceBundle' }
...
  /webhooks:
    get:
      tags: [webhooks]
      summary: Outbound-Webhooks
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Webhook' }
    post:
      tags: [webhooks]
      summary: Webhook anlegen
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Webhook' }
      responses:
        '201': { description: created }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    TenantId:
      name: tenantId
      in: path
      required: true
      schema:
        type: string
        pattern: '^tnt_[a-z0-9]{4,32}$'
  schemas:
...
    Tenant:
      type: object
      properties:
        id: { type: string, pattern: '^tnt_[a-z0-9]{4,32}$' }
        slug: { type: string }
        display_name: { type: string }
        brand:
          type: string
          enum: [ben-e-fit, medialine, ki-guru, custom]
        keycloak_realm: { type: string }
        data_classification_default:
          type: string
          enum: [public, internal, confidential, pii]
        external_llm_allowed: { type: boolean, default: false }
        created_at: { type: string, format: date-time }
      required: [id, slug, display_name, brand, keycloak_realm]
    TenantCreate:
      type: object
      properties:
        slug: { type: string }
        display_name: { type: string }
        brand: { type: string, enum: [ben-e-fit, medialine, ki-guru, custom] }
        admin_email: { type: string, format: email }
        external_llm_allowed: { type: boolean, default: false }
      required: [slug, display_name, brand, admin_email]
    TenantUpdate:
      type: object
      properties:
        display_name: { type: string }
        brand: { type: string, enum: [ben-e-fit, medialine, ki-guru, custom] }
        external_llm_allowed: { type: boolean }
    Source:
      type: object
      properties:
        id: { type: string }
        type: { type: string, enum: [airbyte, upload, rest, crawl4ai, nimble] }
        name: { type: string }
        config: { type: object }
        last_sync: { type: string, format: date-time, nullable: true }
        status: { type: string, enum: [healthy, degraded, failed, unknown] }
      required: [id, type, name]
    SourceCreate:
      type: object
      properties:
        type: { type: string }
        name: { type: string }
        config: { type: object }
      required: [type, name]
...
    DatasetProfile:
      type: object
      properties:
        dataset_id: { type: string }
        columns:
          type: array
          items:
            type: object
            properties:
              name: { type: string }
              type: { type: string }
              null_ratio: { type: number }
              distinct_count: { type: integer }
              sample_values: { type: array, items: { type: string } }
              detected_pii: { type: boolean }
      required: [dataset_id, columns]
...
    MetricDefinitionCreate:
      allOf:
        - $ref: '#/components/schemas/MetricDefinition'
...
    DiscoveryRunRequest:
      type: object
      properties:
        dataset_ids:
          type: array
          items: { type: string }
        scope_hint: { type: string, description: 'Domain hint, e.g. finance' }
      required: [dataset_ids]
    PromoteRequest:
      type: object
      properties:
        approver: { type: string }
        approval_note: { type: string }
        target_metric_name: { type: string }
      required: [approver]
...
    AgentRunRef:
      type: object
      properties:
        run_id: { type: string }
        status: { type: string }
        trace_url: { type: string }
      required: [run_id, status]
...